SP10 Framework: GenericItemSerializer<T> unit test
add a test to ensure sub class types are deserialized when they're assigned to properties of their super class. also inherited public properties are included.

+++ src/Sphere10.Framework/Serialization/GenericItemSerializerOfT.cs
@@ -182,7 +182,7 @@ namespace Sphere10.Framework {
 
 			if (propertyType == typeof(CircularReference)) {
 				var type = ReadTypeHeader(reader);
-				var index = (int)DeserializePrimitive(typeof(int), reader);
+				var index = (ushort)DeserializePrimitive(typeof(ushort), reader);
 				return context.GetObjectByIndex(type, index);
 			}
